+++ internal/app/wwctl/configure/nfs/root.go
@@ -13,9 +13,6 @@ var (
 	}
 )
 
-func init() {
-}
-
 // GetRootCommand returns the root cobra.Command for the application.
 func GetCommand() *cobra.Command {
 	return baseCmd