+++ internal/app/wwctl/genconf/warewulfconf/print/root.go
@@ -14,9 +14,6 @@ var (
 	}
 )
 
-func init() {
-}
-
 func GetCommand() *cobra.Command {
 	return baseCmd
 }