+++ internal/app/wwctl/overlay/create/root.go
@@ -16,9 +16,6 @@ var (
 	}
 )
 
-func init() {
-}
-
 // GetRootCommand returns the root cobra.Command for the application.
 func GetCommand() *cobra.Command {
 	return baseCmd